pklaus.processes.limit.timelimit - try/except is what's needed here
If we catch the signal, it will also not be handed through to
our child process. That's not what we want.
Instead, the KeyboardInterrupt exception that subprocess.run()
throws should just be ignored (while still handed over to the child).
The only output from terminating or killing after the timeout
should be from the child process.